backend/requirements.txt
# Core Django
Django==5.0.2
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Environment and Configuration
python-decouple==3.8
python-dotenv==1.0.0

# Internationalization
django-modeltranslation==0.18.12
django-parler==2.4

# CORS and Security
django-cors-headers==4.3.1

# Filtering and Search
django-filter==23.5

# API Documentation
drf-spectacular==0.27.0

# Image Processing
Pillow==10.1.0

# HTTP Requests
requests==2.31.0

# Date/Time
python-dateutil==2.8.2

# Currency and Number Formatting
babel==2.14.0

# Development Tools
django-debug-toolbar==4.2.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Monitoring and Logging
sentry-sdk==1.38.0

# Caching (for Redis)
redis==5.0.1
django-redis==5.4.0

# Celery (for background tasks)
celery==5.3.4

# Email
django-anymail==10.2

# File Upload
django-storages==1.14.2

# Validation
django-validators==0.4.2

# Utilities
python-slugify==8.0.1 